move pooling types to separate file
